split candlestick e2e tests into separate file
Original candlestick.spec.ts (panning only) restored unchanged.
New candlestick-panels.spec.ts with renders, tooltip, legend,
panel options, and data links tests against a dedicated dashboard.
